backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# WebSocket Support
websockets==12.0
websocket-client==1.5.1
python-socketio==5.10.0

# Data Processing
pandas==2.1.3
numpy==1.26.2

# Technical Analysis (Pure Python - No compilation needed)
ta==0.11.0

# Historical Data Sources (for testing/backtesting)
yfinance==0.2.32
nsepy==0.8

# Kotak Neo Official SDK (optional - experimental)
git+https://github.com/Kotak-Neo/kotak-neo-api.git
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP Requests
requests==2.31.0
httpx==0.25.2

# Caching (optional but recommended)
redis==5.0.1

# Logging & Monitoring
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Utilities
pytz==2023.3
python-dateutil==2.8.2
